make CodeSearchImpl definition local
